+++ src/components/HomeComponent.tsx
@@ -1,16 +1,12 @@
-import React, { useState, useMemo } from 'react';
-import LoginContext from './themes';
-import LoginProvider from './google_login/loginProvider';
-
-// import LoginComponent from './google_login/loginComponent';
-// import LogoutComponent from './google_login/logoutComponent';
-// eslint-disable-next-line func-names
+import React from 'react';
+// import GoogleLoginContext from '../store/GoogleLoginContext';
+// import LoginProvider from './google_login/GoogleLogin';
 
 function HomeComponent({ location }: { location: any }) {
-  const [loginState, setLoginState] = useState(
-    localStorage.getItem('loginstate') === 'true',
-  );
-  const value = useMemo(() => ({ loginState, setLoginState }), [loginState]);
+  // const [loginState, setLoginState] = useState(
+  //   localStorage.getItem('loginstate') === 'true',
+  // );
+  // const value = useMemo(() => ({ loginState, setLoginState }), [loginState]);
   return (
     <h1>
       Content for
@@ -18,10 +14,10 @@ function HomeComponent({ location }: { location: any }) {
       ${location.pathname}
       "`}
       TBD
-      <LoginContext.Provider value={value}>
-        <LoginProvider />
+      {/* <GoogleLoginContext.Provider value={value}>
+      <LoginProvider />
         {loginState ? 'Logged In: True' : 'Logged In: False'}
-      </LoginContext.Provider>
+      </GoogleLoginContext.Provider> */}
     </h1>
   );
 }
