+++ src/components/google_login/loginComponent.tsx
@@ -1,17 +1,17 @@
-import React, { useContext, useEffect } from 'react';
+import React, { useContext } from 'react';
 import GoogleLogin from 'react-google-login';
-import LoginContext from '../themes';
+import GoogleLoginContext from '../../store/GoogleLoginContext';
 
 function LoginComponent() {
-  const { loginState, setLoginState } = useContext(LoginContext);
+  const { setLoginState } = useContext(GoogleLoginContext);
 
   const handleSuccess = () => {
     setLoginState(true);
   };
 
-  useEffect(() => {
-    localStorage.setItem('loginstate', loginState);
-  }, [loginState]);
+  // useEffect(() => {
+  //   localStorage.setItem('loginstate', loginState);
+  // }, [loginState]);
 
   const handleFailure = () => {
     console.log('login failed');
