--- src/components/themes.tsx
@@ -1,8 +1,0 @@
-import { createContext } from 'react';
-
-const defaultvalue: { [key: string]: any } = {
-  loginState: localStorage.getItem('loginstate') === 'true',
-  setLoginState: () => {},
-};
-const LoginContext = createContext(defaultvalue);
-export default LoginContext;
