+++ src/index.tsx
@@ -1,6 +1,7 @@
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { ThemeProvider } from 'styled-components';
+import { GoogleLoginContextProvider } from './store/GoogleLoginContext';
 import theme from './theme/styled-components/theme';
 import GlobalStyle from './theme/styled-components/global.css';
 import './theme/bootstrap/custom.scss';
@@ -11,7 +12,9 @@ ReactDOM.render(
   <ThemeProvider theme={theme}>
     <BrowserRouter>
       <GlobalStyle />
-      <App />
+      <GoogleLoginContextProvider>
+        <App />
+      </GoogleLoginContextProvider>
     </BrowserRouter>
   </ThemeProvider>,
   document.getElementById('root'),
